Shell script to run doxygen
A shell script to run doxygen to produce API docs. There are 3 different options that can be modified in the script – the one used here is `doxygen-awesome` base CSS style. Requires that `doxygen` be installed:
- https://www.doxygen.nl/
- For Macs: https://formulae.brew.sh/formula/doxygen

(Really I forgot to mention this in the first commit of the doxygen API docs `94d18d7`)